requirements.txt
# --- PyTorch / audio backend ---
torch==2.8.0
torchaudio==2.8.0
torchcodec==0.7.0

# --- pyannote ---
pyannote.audio==3.4.0
omegaconf==2.3.0

# --- audio recording ---
pydub==0.25.1
soundfile
numpy==1.26.4
scipy>=1.9.0
librosa>=0.9.2

# --- Transcription ---
faster-whisper

# --- UI & Utilities ---
streamlit==1.37.0
python-dotenv
sounddevice
huggingface_hub==0.19.4
